requirements.txt
Flask
Flask-JWT-Extended
elasticsearch
PyJWT
python-dotenv
cryptography
telethon
pymongo